rename command bus package
