+++ server/tech-blog-app/src/main/kotlin/com/gmurray/tech/blog/author/application/port/in/CreateAuthorCommandHandler.kt
@@ -3,7 +3,7 @@ package com.gmurray.tech.blog.author.application.port.`in`
 import com.gmurray.tech.blog.author.application.port.out.CreateAuthorPort
 import com.gmurray.tech.blog.author.domain.AuthorId
 import com.gmurray.tech.blog.shared.application.annotation.SyncCommandHandler
-import com.gmurray.tech.blog.shared.application.service.CommandHandler
+import com.gmurray.tech.blog.shared.application.command.CommandHandler
 
 @SyncCommandHandler
 class CreateAuthorCommandHandler(private val createAuthorPort: CreateAuthorPort) :
