+++ server/tech-blog-app/src/main/kotlin/com/gmurray/tech/blog/post/application/port/in/CreateBlogPostCommand.kt
@@ -2,7 +2,7 @@ package com.gmurray.tech.blog.post.application.port.`in`
 
 import com.gmurray.tech.blog.post.domain.Categories
 import com.gmurray.tech.blog.post.domain.PostId
-import com.gmurray.tech.blog.shared.application.service.Command
+import com.gmurray.tech.blog.shared.application.command.Command
 
 class CreateBlogPostCommand(
     val authorId: Long,
