+++ server/tech-blog-app/src/main/kotlin/com/gmurray/tech/blog/post/application/port/in/CreateBlogPostCommandHandler.kt
@@ -3,7 +3,7 @@ package com.gmurray.tech.blog.post.application.port.`in`
 import com.gmurray.tech.blog.post.application.port.out.CreateBlogPostPort
 import com.gmurray.tech.blog.post.domain.PostId
 import com.gmurray.tech.blog.shared.application.annotation.SyncCommandHandler
-import com.gmurray.tech.blog.shared.application.service.CommandHandler
+import com.gmurray.tech.blog.shared.application.command.CommandHandler
 
 @SyncCommandHandler
 class CreateBlogPostCommandHandler(private val createBlogPostPort: CreateBlogPostPort) :
