+++ server/tech-blog-app/src/main/kotlin/com/gmurray/tech/blog/shared/application/command/Command.kt
@@ -1,0 +1,3 @@
+package com.gmurray.tech.blog.shared.application.command
+
+interface Command<R>
