+++ server/tech-blog-app/src/main/kotlin/com/gmurray/tech/blog/shared/application/command/CommandHandler.kt
@@ -1,4 +1,4 @@
-package com.gmurray.tech.blog.shared.application.service
+package com.gmurray.tech.blog.shared.application.command
 
 interface CommandHandler<R, C : Command<R>> {
