--- server/tech-blog-app/src/main/kotlin/com/gmurray/tech/blog/shared/application/service/Command.kt
@@ -1,3 +1,0 @@
-package com.gmurray.tech.blog.shared.application.service
-
-interface Command<R>
